Move settings to separate UI and logic components.
